refactor: delete controlmap/ pipeline, fold harvest+ipad_push into slmkiii/
Removes the static-template compile_mapping pipeline (~3000 LoC + ~1100 LoC
tests). The live runtime in scripts/slmk_aum_controller.py never used it;
keeping a parallel half-broken pipeline was the source of the two-data-models
problem (controlmap.model.Binding vs the flat Binding in the runtime).

Removed modules: model, spec_loader, strategy, paging, cc_alloc, value_format,
bridge, emitters/, controllers/, plugins/registry, plugins/, __init__.

Surviving useful pieces moved into slmkiii/:
  - controlmap/plugins/harvest.py -> slmkiii/harvest.py
  - controlmap/ipad_push.py -> slmkiii/ipad_push.py
  - controlmap/plugins/data/{animoog_z,ua_battalion}.json -> slmkiii/data/plugins/
  - controlmap/controllers/data/slmkiii.json -> slmkiii/data/controllers/

Tests: 109 passing (down from 234 — 125 tests removed with the dead pipeline).
