+++ slmkiii/data/plugins/animoog_z.json
